Fix docs/warnings
Review at http://gwt-code-reviews.appspot.com/1380805

Review by: jat@google.com

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@9873 8db76d5a-ed1c-0410-87a9-c151d255dfc7

+++ user/src/com/google/gwt/autobean/server/impl/TypeUtils.java
@@ -50,7 +50,7 @@ public class TypeUtils {
     temp.put(char.class, (char) 0);
     temp.put(double.class, (double) 0);
     temp.put(float.class, (float) 0);
-    temp.put(int.class, (int) 0);
+    temp.put(int.class, 0);
     temp.put(long.class, (long) 0);
     temp.put(short.class, (short) 0);
     temp.put(void.class, null);
